Update Viewer Program.py
Greatly increased the idle-running lifespan of the program on my raspberry pi by eliminating all the needless print statements.

Also added debug variables at the top to control the flow speed of the game.

Also made it so that a red rectangle appears around newly placed stones for 1.5s. (If you don't like it, set highlightTime to 0). This is because when the board gets kinda full, it's hard to see where stones are placed, especially white stones.